Error de conversión de doble a entero #46

+++ src/main/java/es/ubu/lsi/ubumonitor/clustering/algorithm/FuzzyKMeans.java
@@ -23,7 +23,7 @@ public class FuzzyKMeans extends Algorithm {
 	@Override
 	public <T extends Clusterable> Clusterer<T> getClusterer() {
 		int k = getParameters().getValue(ClusteringParameter.NUM_CLUSTER);
-		int fuzziness = getParameters().getValue(ClusteringParameter.FUZZINESS);
+		double fuzziness = getParameters().getValue(ClusteringParameter.FUZZINESS);
 		int max = getParameters().getValue(ClusteringParameter.MAX_ITERATIONS);
 		Distance distance = getParameters().getValue(ClusteringParameter.DISTANCE_TYPE);
